requirements.txt
numpy
pandas
matplotlib
keras
tensorflow
Pillow
ipython
